hdl: nios: devices_rfic: reorg queue-related fns
Move queue-related stuff into devices_rfic_queue.h
